back/pyproject.toml
[tool.poetry]
name = "ecsite"
version = "0.1.0"
description = ""
authors = ["tentaShiratori <tenta.shiratori@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
uritemplate = "^4.1.1"
pyyaml = "^6.0"
django-cors-headers = "^3.14.0"
gunicorn = "^20.1.0"
django-extensions = "^3.2.1"
django-environ = "^0.10.0"
requests = "^2.28.2"
boto3 = "^1.26.98"
boto3-stubs = {extras = ["dynamodb", "s3"], version = "^1.26.98"}
mysqlclient = "^2.1.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pyright = "^1.1.299"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
